ft_substr: Clean up

+++ include/libft.h
@@ -42,7 +42,7 @@ char			*ft_strchr(const char *s, int c);
 char			*ft_strrchr(const char *s, int c);
 char			*ft_strnstr(const char *str, const char *find, size_t n);
 char			*ft_strdup(const char *str);
-char			*ft_substr(const char *s, unsigned int start, size_t len);
+char			*ft_substr(const char *s, size_t start, size_t len);
 char			*ft_strjoin(const char *s1, const char *s2);
 char			*ft_strtrim(const char *str, const char *set);
 char			*ft_strmapi(const char *s, char (*f)(unsigned int, char));
